+++ server/TWS Foundation/TWS Customer/Services/Interfaces/IDriversExternalsService.cs
@@ -6,4 +6,6 @@ using TWS_Business.Sets;
 namespace TWS_Customer.Services.Interfaces;
 public interface IDriversExternalsService {
     Task<SetViewOut<DriverExternal>> View(SetViewOptions<DriverExternal> Options);
+    Task<SetBatchOut<DriverExternal>> Create(DriverExternal[] Driver);
+    Task<RecordUpdateOut<DriverExternal>> Update(DriverExternal Driver);
 }